feat(client): C-1 add ability to login on client.

+++ Faw.Models.Domain/PlayerInfo.cs
@@ -3,6 +3,6 @@
     public class PlayerInfo : BaseEntity
     {
         public int Level { get; set; }
-        public float ExpirienceAmount { get; set; }
+        public decimal ExpirienceAmount { get; set; }
     }
 }
